pyproject.toml
[tool.poetry]
name = "rencons-cli"
version = "0.1.0"
description = ""
authors = ["Nazhmutdin <gumuevnazhmutdin1248@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
naks-library = {git = "https://github.com/Nazhmutdin/naks_library.git"}
click = "^8.1.7"
requests = "^2.32.3"
lxml = "^5.3.0"
pydantic = "^2.9.2"
rich = "^13.9.2"
openpyxl = "^3.1.5"
dishka = "^1.4.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
requests-mock = "^1.12.1"
python-jose = "^3.3.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
